Some KMM fixes
