+++ .gitignore
@@ -12,5 +12,4 @@ captures
 .cxx
 local.properties
 xcuserdata
-course_adapter/src/jvmTest/kotlin/Test.kt
 /.kotlin
