--- .idea/.gitignore
@@ -1,4 +1,0 @@
-# Default ignored files
-/workspace.xml
-/misc.xml# GitHub Copilot persisted chat sessions
-/copilot/chatSessions
