--- .idea/libraries/com_google_code_gson_gson_2_8_6.xml
@@ -1,10 +1,0 @@
-<component name="libraryTable">
-  <library name="com.google.code.gson:gson:2.8.6" type="repository">
-    <properties maven-id="com.google.code.gson:gson:2.8.6" />
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/com/google/code/gson/gson/2.8.6/gson-2.8.6.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES />
-  </library>
-</component>
