--- .idea/libraries/kotlin_csv_jvm_0_11_0.xml
@@ -1,17 +1,0 @@
-<component name="libraryTable">
-  <library name="kotlin-csv-jvm-0.11.0" type="repository">
-    <properties maven-id="com.github.doyaaaaaken:kotlin-csv-jvm:0.11.0" />
-    <CLASSES>
-      <root url="jar://$APPLICATION_HOME_DIR$/jbr/bin/Users/yzune/.m2/repository/org/jetbrains/annotations/13.0/kotlin-csv-jvm-0.11.0.jar!/" />
-      <root url="jar://$APPLICATION_HOME_DIR$/jbr/bin/Users/yzune/.m2/repository/org/jetbrains/annotations/13.0/kotlin-stdlib-jdk8-1.3.72.jar!/" />
-      <root url="jar://$APPLICATION_HOME_DIR$/jbr/bin/Users/yzune/.m2/repository/org/jetbrains/annotations/13.0/kotlin-stdlib-1.3.72.jar!/" />
-      <root url="jar://$APPLICATION_HOME_DIR$/jbr/bin/Users/yzune/.m2/repository/org/jetbrains/annotations/13.0/annotations-13.0.jar!/" />
-      <root url="jar://$APPLICATION_HOME_DIR$/jbr/bin/Users/yzune/.m2/repository/org/jetbrains/annotations/13.0/kotlin-stdlib-jdk7-1.3.72.jar!/" />
-      <root url="jar://$APPLICATION_HOME_DIR$/jbr/bin/Users/yzune/.m2/repository/org/jetbrains/annotations/13.0/kotlin-logging-1.7.9.jar!/" />
-      <root url="jar://$APPLICATION_HOME_DIR$/jbr/bin/Users/yzune/.m2/repository/org/jetbrains/annotations/13.0/slf4j-api-1.7.29.jar!/" />
-      <root url="jar://$APPLICATION_HOME_DIR$/jbr/bin/Users/yzune/.m2/repository/org/jetbrains/annotations/13.0/kotlin-stdlib-common-1.3.72.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES />
-  </library>
-</component>
