--- .idea/libraries/org_jsoup_jsoup_1_12_1.xml
@@ -1,12 +1,0 @@
-<component name="libraryTable">
-  <library name="org.jsoup:jsoup:1.12.1" type="repository">
-    <properties maven-id="org.jsoup:jsoup:1.12.1" />
-    <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/jsoup/jsoup/1.12.1/jsoup-1.12.1.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/jsoup/jsoup/1.12.1/jsoup-1.12.1-sources.jar!/" />
-    </SOURCES>
-  </library>
-</component>
