+++ course_adapter/src/commonMain/kotlin/utils/JsonUtils.kt
@@ -2,10 +2,9 @@ package utils
 
 import kotlinx.serialization.json.Json
 
-val jsonUtil by lazy {
+val jsonUtil =
     Json {
         isLenient = true
         ignoreUnknownKeys = true
         explicitNulls = false
     }
-}
